feat: refactor application structure and included all timeline for my applicaiton.
