+++ src/modules/my-application/my-application.schema.ts
@@ -1,0 +1,2 @@
+export * from "./schema/my-application.request.schema";
+export * from "./schema/my-application.response.schema";
