+++ src/modules/my-application/schema/my-application.response.schema.ts
@@ -26,50 +26,6 @@ export const myApplicationStatusGroupSchema = z
   ])
   .openapi("MyApplicationStatusGroup");
 
-export const getMyApplicationsQuerySchema = z
-  .object({
-    type: z.enum(["all", "volunteer", "projects"]).default("all"),
-    filter: z
-      .enum(["all", "pending", "approved", "active", "completed", "archived"])
-      .default("all"),
-  })
-  .openapi("GetMyApplicationsQuery");
-
-export const myApplicationSourceParamSchema = z
-  .enum(["volunteer", "projects"])
-  .openapi("MyApplicationSourceParam");
-
-export const myApplicationStatusActionSchema = z
-  .enum(["confirm", "decline", "withdraw"])
-  .openapi("MyApplicationStatusAction");
-
-export const myApplicationArchiveActionSchema = z
-  .enum(["archive", "unarchive"])
-  .openapi("MyApplicationArchiveAction");
-
-export const changeMyApplicationStatusParamSchema = z
-  .object({
-    sourceType: myApplicationSourceParamSchema,
-    applicationId: z.string().uuid(),
-    statusAction: myApplicationStatusActionSchema,
-  })
-  .openapi("ChangeMyApplicationStatusParam");
-
-export const changeMyApplicationArchiveParamSchema = z
-  .object({
-    sourceType: myApplicationSourceParamSchema,
-    applicationId: z.string().uuid(),
-    archiveAction: myApplicationArchiveActionSchema,
-  })
-  .openapi("ChangeMyApplicationArchiveParam");
-
-export const getMyApplicationDetailParamSchema = z
-  .object({
-    sourceType: myApplicationSourceParamSchema,
-    applicationId: z.string().uuid(),
-  })
-  .openapi("GetMyApplicationDetailParam");
-
 export const myApplicationItemSchema = z
   .object({
     id: z.string(),
@@ -108,7 +64,12 @@ export const myApplicationsResponseSchema = z
 const myApplicationTimelineSchema = z
   .object({
     submitted: z.string().nullable(),
+    underReview: z.string().nullable(),
     passed: z.string().nullable(),
+    declined: z.object({
+      at: z.string().nullable(),
+      by: z.enum(["POSTER", "APPLICANT"]).nullable(),
+    }),
     confirmed: z.string().nullable(),
     completed: z.string().nullable(),
   })
@@ -186,16 +147,3 @@ export const myApplicationsErrorResponseSchema = z
     error: z.string(),
   })
   .openapi("MyApplicationsErrorResponse");
-
-export type GetMyApplicationsQuery = z.infer<
-  typeof getMyApplicationsQuerySchema
->;
-export type GetMyApplicationDetailParam = z.infer<
-  typeof getMyApplicationDetailParamSchema
->;
-export type ChangeMyApplicationStatusParam = z.infer<
-  typeof changeMyApplicationStatusParamSchema
->;
-export type ChangeMyApplicationArchiveParam = z.infer<
-  typeof changeMyApplicationArchiveParamSchema
->;
